pyproject.toml
[project]
name = "threat-intel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ollama>=0.6.1",
    "pandas>=3.0.2",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "rich>=14.3.4",
    "streamlit>=1.56.0",
    "typer>=0.24.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
